+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * fix parsing complex with 0 inexact imaginary part
 * fix option -t --trace in Node REPL
 * fix `eqv?` on exact and inexact numbers according to R7RS spec
+* fix `exact->inexact` on complex numbers
 
 ## 1.0.0-beta.11
 ### Breaking